LPS-137856 Check if layout is null before trying to delete

+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/model/listener/LayoutModelListener.java
@@ -34,6 +34,10 @@ public class LayoutModelListener extends BaseModelListener<Layout> {
 	@Override
 	public void onAfterRemove(Layout layout) throws ModelListenerException {
 		try {
+			if (layout == null) {
+				return;
+			}
+
 			_segmentsExperienceLocalService.deleteSegmentsExperiences(
 				layout.getGroupId(),
 				_classNameLocalService.getClassNameId(Layout.class),
